added media upload

+++ src/app/components/popover/popover.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from "@angular/core";
-import { AlertController } from "@ionic/angular";
 
 import { Message } from "src/app/interfaces/chat.model";
+import { AlertService } from "src/app/services/alert.service";
 import { ConversationService } from "src/app/services/conversation.service";
 import { PopoverService } from "src/app/services/popover.service";
 
@@ -19,7 +19,7 @@ export class PopoverComponent implements OnInit {
   constructor(
     private popoverService: PopoverService,
     private conversastionService: ConversationService,
-    private alrtController: AlertController
+    private alrtService: AlertService
   ) {}
 
   ngOnInit() {}
@@ -27,43 +27,24 @@ export class PopoverComponent implements OnInit {
     this.popoverService.dismissPopover();
     const result = await this.popoverService.getPopoverResult();
     const data: Message = { ...Object.values<Message>(result)[0] };
-    console.log(data);
     this.message = [data];
   }
 
   async deleteMessage() {
     await this.dismissPopover();
-    console.log("Message", this.message);
-    const alert = await this.alrtController.create({
-      cssClass: "my-custom-class",
-      header: `Confirm`,
-      message: "All the messages will be deleted forever!!!",
-      buttons: [
-        {
-          text: "Cancel",
-          role: "cancel",
-          cssClass: "secondary",
-          handler: (blah) => {
-            console.log("Confirm Cancel: blah");
-          },
-        },
-        {
-          text: "Okay",
-          handler: () => {
-            console.log("Confirm Okay");
-
-            this.conversastionService.deleteMessages(
-              this.message,
-              this.message[0].conversationId
-            );
-          },
-        },
-      ],
-    });
-    await alert.present();
+    const alertResult = await this.alrtService.presentAlertConfirm(
+      `Confirm`,
+      "All the messages will be deleted forever!!!"
+    );
+    if (alertResult === "continue") {
+      this.conversastionService.deleteMessages(
+        this.message,
+        this.message[0].conversationId
+      );
+    } else return;
   }
 
-  public async replyMessage() {
-    await this.dismissPopover();
+  async replyMessage() {
+    return await this.dismissPopover();
   }
 }
